Complete NeoSwift parity and enable tests

+++ include/neoc/protocol/core/response/neo_get_nep11_balances.h
@@ -29,7 +29,7 @@ extern "C" {
 typedef struct {
     char *token_id;                     /**< Unique token identifier */
     char *amount;                       /**< Token amount as string */
-    int last_updated_block;             /**< Block when last updated */
+    int64_t last_updated_block;         /**< Block when last updated */
 } neoc_nep11_token_t;
 
 /**
