+++ include/neoc/protocol/core/response/neo_get_nep17_transfers.h
@@ -28,10 +28,10 @@ extern "C" {
  * timestamps, addresses, amounts, and blockchain references.
  */
 typedef struct {
-    int timestamp;                      /**< Transfer timestamp */
+    int64_t timestamp;                  /**< Transfer timestamp */
     neoc_hash160_t asset_hash;          /**< Token contract hash */
     char *transfer_address;             /**< Transfer counterpart address */
-    int amount;                         /**< Transfer amount as integer */
+    int64_t amount;                     /**< Transfer amount as integer */
     int block_index;                    /**< Block index where transfer occurred */
     int transfer_notify_index;          /**< Transfer notification index */
     neoc_hash256_t tx_hash;             /**< Transaction hash */
@@ -78,10 +78,10 @@ typedef struct {
  * @return NEOC_SUCCESS on success, error code on failure
  */
 neoc_error_t neoc_nep17_transfer_create(
-    int timestamp,
+    int64_t timestamp,
     const neoc_hash160_t *asset_hash,
     const char *transfer_address,
-    int amount,
+    int64_t amount,
     int block_index,
     int transfer_notify_index,
     const neoc_hash256_t *tx_hash,
